+++ src/main/java/moldmod/integration/jade/JadePlugin.java
@@ -1,0 +1,14 @@
+package moldmod.integration.jade;
+
+import snownee.jade.api.IWailaClientRegistration;
+import snownee.jade.api.IWailaPlugin;
+import snownee.jade.api.WailaPlugin;
+import net.minecraft.block.Block;
+
+@WailaPlugin
+public class JadePlugin implements IWailaPlugin {
+    @Override
+    public void registerClient(IWailaClientRegistration registration) {
+        registration.registerBlockComponent(MoldyBlockProvider.INSTANCE, Block.class);
+    }
+}